Move response-related methods to FeedbackResponseLogic
